Enable OPD RegisterProvider source data configurability
This means that you can have different columns names for guardian details & client details, different register-type column names or not have them at all. All you need to do is implement the interface and provide the class to the configuration. All this means that you can define the kind of data that is shown on the client row in the OPD register by configuring this as you wish. However, the base data type i.e. client first name, middle name, last name, gender, year etc. and layout cannot be changed using this class

- This commit also provides a default implementation that assumes how certain column names are named and can be fetched from the column map that is generated by the CommonPersonObjectClient